almost implemented pmalloc

+++ miniOS/include/miniOS.h
@@ -45,7 +45,7 @@ extern int yield(void);
 /*
  * Heap memory manager
  */
-extern void *hm_malloc(long int size);
+extern void *hm_malloc(size_t size);
 extern void *hm_realloc(void* ptr, long int size);
 extern void hm_free(void *ptr);
